cassandra/requirements.txt
cassandra-driver==3.29.1
python-dotenv==1.0.1